refactor: chain entry validation into contract interface
